Move stocktake lines insert to respective folder/files
