+++ server/service/src/stocktake_line/insert/mod.rs
@@ -1,19 +1,13 @@
+mod validate;
+use validate::{validate, GenerateResult};
+mod generate;
+use generate::generate;
+
 use chrono::NaiveDate;
-use repository::{EqualFilter, ItemFilter, ItemRepository};
-use repository::{
-    RepositoryError, StockLine, StocktakeLine, StocktakeLineFilter, StocktakeLineRepository,
-    StocktakeLineRow, StocktakeLineRowRepository, StorageConnection,
-};
+use repository::{RepositoryError, StockLine, StocktakeLine, StocktakeLineRowRepository};
 
-use crate::common_stock::{check_stock_line_exists, CommonStockLineError};
-use crate::stocktake::{check_stocktake_exist, check_stocktake_not_finalised};
-use crate::validate::check_store_id_matches;
-use crate::{check_location_exists, NullableUpdate};
+use crate::NullableUpdate;
 use crate::{service_provider::ServiceContext, stocktake_line::query::get_stocktake_line};
-
-use super::validate::{
-    check_active_adjustment_reasons, check_reason_is_valid, check_stock_line_reduced_below_zero,
-};
 
 #[derive(Default, Debug, Clone)]
 pub struct InsertStocktakeLine {
@@ -51,236 +45,6 @@ pub enum InsertStocktakeLineError {
     AdjustmentReasonNotProvided,
     AdjustmentReasonNotValid,
     StockLineReducedBelowZero(StockLine),
-}
-
-fn check_stocktake_line_does_not_exist(
-    connection: &StorageConnection,
-    id: &str,
-) -> Result<bool, RepositoryError> {
-    let count = StocktakeLineRepository::new(connection).count(
-        Some(StocktakeLineFilter::new().id(EqualFilter::equal_to(id))),
-        None,
-    )?;
-    Ok(count == 0)
-}
-
-fn check_stock_line_is_unique(
-    connection: &StorageConnection,
-    id: &str,
-    stock_line_id: &str,
-) -> Result<bool, RepositoryError> {
-    let stocktake_lines = StocktakeLineRepository::new(connection).query_by_filter(
-        StocktakeLineFilter::new().stocktake_id(EqualFilter::equal_to(id)),
-        None,
-    )?;
-    let already_has_stock_line = stocktake_lines.iter().find(|line| {
-        if let Some(ref stock_line) = line.stock_line {
-            return stock_line.id == stock_line_id;
-        }
-        false
-    });
-    match already_has_stock_line {
-        Some(_) => Ok(false),
-        None => Ok(true),
-    }
-}
-
-/// If valid it returns the item_id it either from the stock_line or from input.item_id
-fn check_stock_line_xor_item(
-    stock_line: &Option<StockLine>,
-    input: &InsertStocktakeLine,
-) -> Option<String> {
-    if (stock_line.is_none() && input.item_id.is_none())
-        || (stock_line.is_some() && input.item_id.is_some())
-    {
-        return None;
-    }
-
-    // extract item_id
-    if let Some(stock_line) = stock_line {
-        return Some(stock_line.item_row.id.clone());
-    }
-    input.item_id.clone()
-}
-
-pub fn stocktake_reduction_amount(
-    counted_number_of_packs: &Option<f64>,
-    stock_line: &Option<StockLine>,
-) -> f64 {
-    if let (Some(stock_line), Some(counted_number_of_packs)) = (stock_line, counted_number_of_packs)
-    {
-        stock_line.stock_line_row.total_number_of_packs - counted_number_of_packs
-    } else if stock_line.is_none() && counted_number_of_packs.is_some() {
-        -counted_number_of_packs.unwrap_or(0.0)
-    } else {
-        0.0
-    }
-}
-
-pub fn check_item_exists_and_get_item_name(
-    connection: &StorageConnection,
-    store_id: &str,
-    item_id: &str,
-) -> Result<String, InsertStocktakeLineError> {
-    let item = ItemRepository::new(connection)
-        .query_by_filter(
-            ItemFilter::new().id(EqualFilter::equal_to(item_id)),
-            Some(store_id.to_string()),
-        )?
-        .pop()
-        .ok_or(InsertStocktakeLineError::ItemDoesNotExist)?;
-
-    Ok(item.item_row.name)
-}
-
-pub(crate) struct GenerateResult {
-    pub(crate) stock_line: Option<StockLine>,
-    pub(crate) item_id: String,
-    pub(crate) item_name: String,
-}
-
-fn validate(
-    connection: &StorageConnection,
-    store_id: &str,
-    input: &InsertStocktakeLine,
-) -> Result<GenerateResult, InsertStocktakeLineError> {
-    use InsertStocktakeLineError::*;
-
-    let stocktake = match check_stocktake_exist(connection, &input.stocktake_id)? {
-        Some(stocktake) => stocktake,
-        None => return Err(StocktakeDoesNotExist),
-    };
-    if !check_stocktake_not_finalised(&stocktake.status) {
-        return Err(CannotEditFinalised);
-    }
-    if !check_store_id_matches(store_id, &stocktake.store_id) {
-        return Err(InvalidStore);
-    }
-    if !check_stocktake_line_does_not_exist(connection, &input.id)? {
-        return Err(StocktakeLineAlreadyExists);
-    }
-
-    if stocktake.is_locked {
-        return Err(StocktakeIsLocked);
-    }
-
-    let stock_line = if let Some(stock_line_id) = &input.stock_line_id {
-        Some(
-            check_stock_line_exists(connection, store_id, stock_line_id).map_err(
-                |err| match err {
-                    CommonStockLineError::DatabaseError(RepositoryError::NotFound) => {
-                        StockLineDoesNotExist
-                    }
-                    CommonStockLineError::StockLineDoesNotBelongToStore => InvalidStore,
-                    CommonStockLineError::DatabaseError(error) => DatabaseError(error),
-                },
-            )?,
-        )
-    } else {
-        None
-    };
-    if let Some(stock_line) = &stock_line {
-        if !check_stock_line_is_unique(
-            connection,
-            &input.stocktake_id,
-            &stock_line.stock_line_row.id,
-        )? {
-            return Err(StockLineAlreadyExistsInStocktake);
-        }
-    }
-
-    let item_id = check_stock_line_xor_item(&stock_line, input)
-        .ok_or(InsertStocktakeLineError::StockLineXOrItem)?;
-
-    let item_name = if input.item_id.is_some() {
-        check_item_exists_and_get_item_name(connection, store_id, &item_id)?
-    } else {
-        stock_line.as_ref().unwrap().item_row.name.clone()
-    };
-
-    if !check_location_exists(connection, store_id, &input.location)? {
-        return Err(LocationDoesNotExist);
-    }
-
-    let stocktake_reduction_amount =
-        stocktake_reduction_amount(&input.counted_number_of_packs, &stock_line);
-    if check_active_adjustment_reasons(connection, stocktake_reduction_amount)?.is_some()
-        && input.inventory_adjustment_reason_id.is_none()
-        && stocktake_reduction_amount != 0.0
-    {
-        return Err(AdjustmentReasonNotProvided);
-    }
-
-    if input.inventory_adjustment_reason_id.is_some()
-        && !check_reason_is_valid(
-            connection,
-            input.inventory_adjustment_reason_id.clone(),
-            stocktake_reduction_amount,
-        )?
-    {
-        return Err(AdjustmentReasonNotValid);
-    }
-
-    if let (Some(counted_number_of_packs), Some(stock_line)) =
-        (input.counted_number_of_packs, stock_line.clone())
-    {
-        if check_stock_line_reduced_below_zero(&stock_line.stock_line_row, &counted_number_of_packs)
-        {
-            return Err(StockLineReducedBelowZero(stock_line));
-        }
-    }
-
-    Ok(GenerateResult {
-        stock_line,
-        item_id,
-        item_name,
-    })
-}
-
-fn generate(
-    stock_line: Option<StockLine>,
-    item_id: String,
-    item_name: String,
-    InsertStocktakeLine {
-        id,
-        stocktake_id,
-        stock_line_id,
-        location,
-        comment,
-        counted_number_of_packs,
-        item_id: _,
-        batch,
-        expiry_date,
-        pack_size,
-        cost_price_per_pack,
-        sell_price_per_pack,
-        note,
-        inventory_adjustment_reason_id,
-    }: InsertStocktakeLine,
-) -> StocktakeLineRow {
-    let snapshot_number_of_packs = if let Some(stock_line) = stock_line {
-        stock_line.stock_line_row.total_number_of_packs
-    } else {
-        0.0
-    };
-    StocktakeLineRow {
-        id,
-        stocktake_id,
-        stock_line_id,
-        location_id: location.map(|l| l.value).unwrap_or_default(),
-        comment,
-        snapshot_number_of_packs,
-        counted_number_of_packs,
-        item_link_id: item_id.to_string(),
-        item_name,
-        batch,
-        expiry_date,
-        pack_size,
-        cost_price_per_pack,
-        sell_price_per_pack,
-        note,
-        inventory_adjustment_reason_id,
-    }
 }
 
 pub fn insert_stocktake_line(
